+++ src/bitstream.h
@@ -6,10 +6,10 @@
 class CBitStream {
 private:
 	FILE* file;
-	unsigned long accumulator; // not yet stored bits
+	uint32_t accumulator; // not yet stored bits
 	int shift; // next bit position
-	unsigned long bits_written, bytes_written; // for statistics
-	long seq_zeros; // count of sequential zeros
+	uint32_t bits_written, bytes_written; // for statistics
+	int32_t seq_zeros; // count of sequential zeros
 
 	void internal_flush(); // drops all the fully filled bytes
 	void drop_zeros(); // flushes all the sequentially coming zero bytes
@@ -23,9 +23,9 @@ public:
 	
 	int init_bit_stream() { return 1; }; // no initialization in current version
 	void flush(); // flushes the last, unfinished byte. MUST be called at the and of processing
-	void write_bits (unsigned long value, int count);
-	unsigned long get_bits_written() { return bits_written; };
-	unsigned long get_bytes_written() { return bytes_written; };
+	void write_bits (uint32_t value, int count);
+	uint32_t get_bits_written() { return bits_written; };
+	uint32_t get_bytes_written() { return bytes_written; };
 };
 
 #endif
