+++ src/snd2acm.h
@@ -1,6 +1,8 @@
 #ifndef _UH_SND2ACM_H
 #define _UH_SND2ACM_H
 
-int ConvertWavAcm(int fh, long maxlen, FILE *foutp, bool wavc_or_acm);
+#include <cstdint>
+
+int ConvertWavAcm(int fh, int32_t maxlen, FILE *foutp, bool wavc_or_acm);
 
 #endif
